+++ src/at.js
@@ -1,4 +1,0 @@
-import { atOr } from './atOr'
-import { setName } from './internal/setName'
-
-export const at = setName(atOr(undefined), 'at')