+++ src/atOr.js
@@ -1,5 +1,0 @@
-import { curry3 } from './curry3'
-import { atOrU } from './internal/atOrU'
-import { setName } from './internal/setName'
-
-export const atOr = setName(curry3(atOrU), 'atOr')