+++ src/internal/nthSatiesfiesU.js
@@ -1,7 +1,7 @@
 import { getIndex } from './getIndex'
 import { hasIndex } from './hasIndex'
 
-export const atSatisfiesU = (fn, index, array) => {
+export const nthSatisfiesU = (fn, index, array) => {
   index = getIndex(index, array)
   return hasIndex(index, array) && fn(array[index])
 }