+++ src/internal/removeNthU.js
@@ -1,0 +1,3 @@
+import { setNthU } from './setNthU'
+
+export const removeNthU = (index, array) => setNthU(index, undefined, array)