+++ src/modifyAt.js
@@ -1,5 +1,0 @@
-import { curry3 } from './curry3'
-import { modifyAtU } from './internal/modifyAtU'
-import { setName } from './internal/setName'
-
-export const modifyAt = setName(curry3(modifyAtU), 'modifyAt')